pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "studyxiao", email = "studyxiao@163.com"},
]
dependencies = [
    "pypdf>=3.9.1",
    "python-dotenv>=1.0.0",
    "openai>=0.27.8",
    "tenacity>=8.2.2",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.272",
    "black>=23.3.0",
    "bpython>=0.24",
    "rich>=13.4.1",
]

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
fix = true
extend-select = ["I"]
